requirements.txt
attrs==17.4.0
beautifulsoup4==4.6.0
bokeh==0.12.14
boto==2.48.0
boto3==1.5.25
botocore==1.8.39
bz2file==0.98
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
cheroot==6.0.0
CherryPy==14.0.0
colorama==0.3.9
cycler==0.10.0
docutils==0.14
docx==0.2.4
feedparser==5.2.1
gensim==3.3.0
google-api-core==0.1.4
google-auth==1.4.0
google-cloud-bigquery==0.30.0
google-cloud-core==0.28.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
Jinja2==2.10
jmespath==0.9.3
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.2
more-itertools==4.1.0
Morfessor==2.0.3
nltk==3.2.5
numpy==1.14.0
packaging==16.8
pandas==0.22.0
pattern3==3.0.0
pdfminer.six==20170720
pdfminer3k==1.3.1
Pillow==5.0.0
pluggy==0.6.0
ply==3.10
polyglot==16.7.4
portend==2.2
protobuf==3.5.1
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycld2==0.31
pycryptodome==3.4.11
PyICU==2.0.3
pyparsing==2.2.0
pypiwin32==220
pytest==3.4.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
simplejson==3.13.2
six==1.11.0
smart-open==1.5.6
tempora==1.10
tornado==4.5.3
urllib3==1.22
